fix : service login 초안 작성완료
applyInformation, tester review, maker review service 로직 추가
* 아직 serivce 예외처리는 안함

Resolves: #issueNo
See also: #issueNo

+++ src/main/java/io/wisoft/testermatchingplatform/web/dto/request/CreateMissionRequest.java
@@ -21,7 +21,7 @@ public class CreateMissionRequest {
     private final String durationTimeStart;
     private final String durationTimeLimit;
     private final String content;
-    private final int reward;
+    private final long reward;
     private final int limitPerformer;
     @Image(groups = Custom.class)
     private final MultipartFile image;
@@ -50,7 +50,7 @@ public class CreateMissionRequest {
             final String durationTimeStart,
             final String durationTimeLimit,
             final String content,
-            final int reward,
+            final long reward,
             final int limitPerformer,
             final MultipartFile symbolImage
     ) {
